Adjust Weight to u64 (#2181)
* Adjust Weight to u64

* Adjust copy (really should be reading folders...)

* build:interfaces

* EventRecord test expectation update

+++ packages/types/src/interfaces/runtime/definitions.ts
@@ -85,7 +85,7 @@ export default {
     },
     StorageData: 'Bytes',
     ValidatorId: 'AccountId',
-    Weight: 'u32',
+    Weight: 'u64',
     WeightMultiplier: 'Fixed64',
 
     // digest
