+++ packages/types/src/interfaces/runtime/types.ts
@@ -187,7 +187,7 @@ export interface StorageData extends Bytes {}
 export interface ValidatorId extends AccountId {}
 
 /** @name Weight */
-export interface Weight extends u32 {}
+export interface Weight extends u64 {}
 
 /** @name WeightMultiplier */
 export interface WeightMultiplier extends Fixed64 {}
